Move `NetworkStats` to `shared::backend`
It's associated with both connected client on the server and the client
itself.
